@@ -3,19 +3,37 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@material-ui/core": "^4.11.0",
-    "express": "^4.16.3",
-    "react": "^16.5.0",
-    "react-dom": "^16.5.0",
-    "react-scripts": "^4.0.0",
-    "socket.io": "^2.1.1",
-    "socket.io-client": "^2.1.1"
+    "@material-ui/core": "^4.12.4",
+    "cors": "^2.8.5",
+    "express": "^4.21.2",
+    "prop-types": "^15.8.1",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
+    "socket.io": "^4.8.1",
+    "socket.io-client": "^4.8.1"
   },
   "scripts": {
+    "dev": "concurrently \"node server.js\" \"react-scripts start\"",
     "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test --env=jsdom",
-    "eject": "react-scripts eject"
+    "build": "react-scripts build"
   },
-  "proxy": "http://localhost:4869"
+  "proxy": "http://localhost:4869",
+  "devDependencies": {
+    "@types/react": "^16.8.6",
+    "concurrently": "^9.1.2",
+    "react-scripts": "^5.0.1",
+    "typescript": "^3.2.1"
+  },
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
+  }
 }
