@@ -1,85 +1,68 @@
-import React, { Component } from "react";
+import React, { useEffect, useState } from "react";
 import socketIOClient from "socket.io-client";
-import PropTypes from "prop-types";
-import { withStyles } from "@material-ui/core/styles";
+
 import AppBar from "@material-ui/core/AppBar";
-import Tabs from "@material-ui/core/Tabs";
 import Tab from "@material-ui/core/Tab";
-import Typography from "@material-ui/core/Typography";
+import Tabs from "@material-ui/core/Tabs";
 
 import TabContainer from "./components/TabContainer";
 
-const styles = theme => ({
+const styles = {
   root: {
-    flexGrow: 1
+    flexGrow: 1,
   },
   tabToolbar: {
     backgroundColor: "#20c1db",
     width: "80%",
     marginLeft: "10%",
     boxShadow: "none",
-    position: "relative"
+    position: "relative",
   },
   tabLabel: {
     display: "flex",
-    justifyContent: "flex-end"
+    justifyContent: "flex-end",
   },
   tabLabelItem: {
-    fontWeight: "normal"
+    fontWeight: "normal",
   },
-  brand: {
-    position: "absolute",
-    top: "50%",
-    transform: "translateY(-50%)",
-    left: "1em",
-    color: "#fff",
-    fontWeight: "500"
-  }
-});
-
-class App extends Component {
-  state = {
-    value: 0,
-    endpoint: "http://localhost:4869"
-  };
-
-  _handleChange = (event, value) => {
-    this.setState({ value });
-  };
-
-  render() {
-    const { classes } = this.props;
-    const { value, data } = this.state;
-
-    return (
-      <div className={classes.root}>
-        <AppBar position="static" className={classes.tabToolbar}>
-          <Typography className={classes.brand}>S&P/CND</Typography>
-          <Tabs
-            value={value}
-            onChange={this._handleChange}
-            className={classes.tabLabel}
-          >
-            <Tab label="top gainers" className={classes.tabLabelItem} />
-            <Tab label="top losers" className={classes.tabLabelItem} />
-          </Tabs>
-        </AppBar>
-        {value === 0 && <TabContainer data={data} order="desc" />}
-        {value === 1 && <TabContainer data={data} order="asc" />}
-      </div>
-    );
-  }
-
-  componentDidMount() {
-    const socket = socketIOClient(this.state.endpoint);
-    socket.on("change data", data => {
-      this.setState({ data });
-    });
-  }
-}
-
-App.propTypes = {
-  classes: PropTypes.object.isRequired
 };
 
-export default withStyles(styles)(App);
+const App = () => {
+  const [value, setValue] = useState(0);
+  const [data, setData] = useState([]);
+
+  const handleChange = (_, value) => {
+    console.log("numeric", value);
+    setValue(value);
+  };
+
+  useEffect(() => {
+    const socket = socketIOClient("http://localhost:4869", {
+      withCredentials: true,
+      transports: ["websocket", "polling"],
+    });
+
+    socket.on("change data", (data) => {
+      setData(data);
+    });
+
+    return () => {
+      socket.disconnect();
+    };
+  }, []);
+
+  return (
+    <div style={styles.root}>
+      <AppBar position="static" style={styles.tabToolbar}>
+        <Tabs value={value} onChange={handleChange} style={styles.tabLabel}>
+          <Tab label="top gainers" style={styles.tabLabelItem} />
+          <Tab label="top losers" style={styles.tabLabelItem} />
+        </Tabs>
+      </AppBar>
+      {value === 0 && <TabContainer data={data} order="desc" />}
+      {value === 1 && <TabContainer data={data} order="asc" />}
+    </div>
+  );
+};
+
+export default App;
