@@ -7,71 +7,71 @@ const randomNumberInRange = (start, end) => {
   return Math.floor(Math.random() * (1 + end - start)) + start;
 };
 
-const formatNumberWithTwoFloatPoint = floatNumber => {
+export const formatNumberWithTwoFloatPoint = (floatNumber) => {
   const NUMBER_FLOAT_POINT = 2;
   return floatNumber.toFixed(NUMBER_FLOAT_POINT);
 };
 
-const getNewVolume = previousVolume => {
+export const getNewVolume = (previousVolume) => {
   const START = 10;
   const END = 30;
   return previousVolume + randomNumberInRange(START, END);
 };
 
-const getNewPriceFromCurrentPrice = currentPrice => {
+export const getNewPriceFromCurrentPrice = (currentPrice) => {
   const MIN_PERCENT = 95;
   const MAX_PERCENT = 105;
   const rawFloatPrice = parseFloat(
     randomNumberInRange(
       currentPrice * MIN_PERCENT,
-      currentPrice * MAX_PERCENT
-    ) / 100
+      currentPrice * MAX_PERCENT,
+    ) / 100,
   );
   return formatNumberWithTwoFloatPoint(rawFloatPrice);
 };
 
-const modifyPriceAndVolume = rowData => {
+export const modifyPriceAndVolume = (rowData) => {
   return {
     ...rowData,
-    lastestPrice: getNewPriceFromCurrentPrice(rowData.lastestPrice),
-    lastestVolume: getNewVolume(rowData.lastestVolume)
+    latestPrice: getNewPriceFromCurrentPrice(rowData.latestPrice),
+    latestVolume: getNewVolume(rowData.latestVolume),
   };
 };
 
-const modifyValue = rowData => {
+export const modifyValue = (rowData) => {
   return {
     ...rowData,
-    value: parseInt(rowData.lastestPrice * rowData.lastestVolume, 10)
+    value: parseInt(rowData.latestPrice * rowData.latestVolume, 10),
   };
 };
 
-const modifyChange = rowData => {
+export const modifyChange = (rowData) => {
   return {
     ...rowData,
     change: formatNumberWithTwoFloatPoint(
-      rowData.lastestPrice - rowData.originalPrice
-    )
+      rowData.latestPrice - rowData.originalPrice,
+    ),
   };
 };
 
-const modifyChangePercent = rowData => {
+export const modifyChangePercent = (rowData) => {
   return {
     ...rowData,
     "%change": formatNumberWithTwoFloatPoint(
-      (rowData.change * 100) / rowData.originalPrice
-    )
+      (rowData.change * 100) / rowData.originalPrice,
+    ),
   };
 };
 
-const modifyStatus = rowData => {
+export const modifyStatus = (rowData) => {
   return {
     ...rowData,
-    status: rowData.change >= 0 ? "increase" : "decrease"
+    status: rowData.change >= 0 ? "increase" : "decrease",
   };
 };
 
-const compareValuesByProperty = (property, order = "asc") => {
-  return function(a, b) {
+export const compareValuesByProperty = (property, order = "asc") => {
+  return function (a, b) {
     if (!a.hasOwnProperty(property) || !b.hasOwnProperty(property)) {
       return 0;
     }
@@ -89,16 +89,4 @@ const compareValuesByProperty = (property, order = "asc") => {
     }
     return order === "desc" ? comparison * -1 : comparison;
   };
-};
-
-module.exports = {
-  getNewVolume,
-  getNewPriceFromCurrentPrice,
-  formatNumberWithTwoFloatPoint,
-  modifyPriceAndVolume,
-  modifyValue,
-  modifyChange,
-  modifyChangePercent,
-  modifyStatus,
-  compareValuesByProperty
 };
