@@ -1,334 +1,321 @@
-const originalData = [
+export const originalData = [
   {
     code: "CBA.GOW",
     company: "ROYAL BANK OF CANADA",
     originalPrice: 80.72,
-    lastestPrice: 0,
+    latestPrice: 0,
     originalVolume: 168129,
-    lastestVolume: 0,
+    latestVolume: 0,
     change: "0.00",
     "%change": "0.00",
-    status: ""
+    status: "",
   },
   {
     code: "CFD.GOW",
     company: "TORONTO DOMINATION BANK",
     originalPrice: 60,
-    lastestPrice: 0,
+    latestPrice: 0,
     originalVolume: 129380,
-    lastestVolume: 0,
+    latestVolume: 0,
     change: "0.00",
     "%change": "0.00",
-    status: ""
+    status: "",
   },
   {
     code: "CAS.GOW",
     company: "BANK OF NOVA SCOTIA",
     originalPrice: 60.34,
-    lastestPrice: 0,
+    latestPrice: 0,
     originalVolume: 123987,
-    lastestVolume: 0,
+    latestVolume: 0,
     change: "0.00",
     "%change": "0.00",
-    status: ""
+    status: "",
   },
   {
     code: "CRE.GOW",
     company: "BANK OF MOTREAL",
     originalPrice: 52.73,
-    lastestPrice: 0,
+    latestPrice: 0,
     originalVolume: 3862,
-    lastestVolume: 0,
+    latestVolume: 0,
     change: "0.00",
     "%change": "0.00",
-    status: ""
+    status: "",
   },
   {
     code: "CEW.GOW",
     company: "CANADIAN IMPERIAL BANK OF COMMERCE",
     originalPrice: 90.91,
-    lastestPrice: 0,
+    latestPrice: 0,
     originalVolume: 93749,
-    lastestVolume: 0,
+    latestVolume: 0,
     change: "0.00",
     "%change": "0.00",
-    status: ""
+    status: "",
   },
   {
     code: "GFA.GOW",
     company: "COMMONWEALTH BANK OF AUSTRALIA",
     originalPrice: 12.12,
-    lastestPrice: 0,
+    latestPrice: 0,
     originalVolume: 168129,
-    lastestVolume: 0,
+    latestVolume: 0,
     change: "0.00",
     "%change": "0.00",
-    status: ""
+    status: "",
   },
   {
     code: "DFV.GOW",
     company: "BANK OF SINGAPORE",
     originalPrice: 12.11,
-    lastestPrice: 0,
+    latestPrice: 0,
     originalVolume: 218014,
-    lastestVolume: 0,
+    latestVolume: 0,
     change: "0.00",
     "%change": "0.00",
-    status: ""
+    status: "",
   },
   {
     code: "ASV.GOW",
     company: "DBS BANK",
     originalPrice: 4.56,
-    lastestPrice: 0,
+    latestPrice: 0,
     originalVolume: 192731,
-    lastestVolume: 0,
+    latestVolume: 0,
     change: "0.00",
     "%change": "0.00",
-    status: ""
+    status: "",
   },
   {
     code: "AWE.GOW",
     company: "OVERSEA CHINESE BANKING CORPORATION",
     originalPrice: 78.23,
-    lastestPrice: 0,
+    latestPrice: 0,
     originalVolume: 901287,
-    lastestVolume: 0,
+    latestVolume: 0,
     change: "0.00",
     "%change": "0.00",
-    status: ""
+    status: "",
   },
   {
     code: "AWS.GOW",
     company: "UNITED OVERSEAS BANK",
     originalPrice: 76.23,
-    lastestPrice: 0,
+    latestPrice: 0,
     originalVolume: 389109,
-    lastestVolume: 0,
+    latestVolume: 0,
     change: "0.00",
     "%change": "0.00",
-    status: ""
+    status: "",
   },
   {
     code: "ERD.GOW",
     company: "JPMORGAN CHASE",
     originalPrice: 18.29,
-    lastestPrice: 0,
+    latestPrice: 0,
     originalVolume: 927211,
-    lastestVolume: 0,
+    latestVolume: 0,
     change: "0.00",
     "%change": "0.00",
-    status: ""
+    status: "",
   },
   {
     code: "JGM.GOW",
     company: "CITY GROUP",
     originalPrice: 38.76,
-    lastestPrice: 0,
+    latestPrice: 0,
     originalVolume: 378146,
-    lastestVolume: 0,
+    latestVolume: 0,
     change: "0.00",
     "%change": "0.00",
-    status: ""
+    status: "",
   },
   {
     code: "LKK.GOW",
     company: "WELLS FARGO",
     originalPrice: 57.54,
-    lastestPrice: 0,
+    latestPrice: 0,
     originalVolume: 156212,
-    lastestVolume: 0,
+    latestVolume: 0,
     change: "0.00",
     "%change": "0.00",
-    status: ""
+    status: "",
   },
   {
     code: "CVV.GOW",
     company: "U.S BANCORP",
     originalPrice: 54.34,
-    lastestPrice: 0,
+    latestPrice: 0,
     originalVolume: 123675,
-    lastestVolume: 0,
+    latestVolume: 0,
     change: "0.00",
     "%change": "0.00",
-    status: ""
+    status: "",
   },
   {
     code: "DSS.GOW",
     company: "STATE FARM",
     originalPrice: 45.12,
-    lastestPrice: 0,
+    latestPrice: 0,
     originalVolume: 763412,
-    lastestVolume: 0,
+    latestVolume: 0,
     change: "0.00",
     "%change": "0.00",
-    status: ""
+    status: "",
   },
   {
     code: "RTR.GOW",
     company: "NORTHERN TRUST",
     originalPrice: 67.23,
-    lastestPrice: 0,
+    latestPrice: 0,
     originalVolume: 297431,
-    lastestVolume: 0,
+    latestVolume: 0,
     change: "0.00",
     "%change": "0.00",
-    status: ""
+    status: "",
   },
   {
     code: "KEO.GOW",
     company: "KETCORP",
     originalPrice: 98.87,
-    lastestPrice: 0,
+    latestPrice: 0,
     originalVolume: 978127,
-    lastestVolume: 0,
+    latestVolume: 0,
     change: "0.00",
     "%change": "0.00",
-    status: ""
+    status: "",
   },
   {
     code: "KIT.GOW",
     company: "SILICON VALLEY BANK",
     originalPrice: 35.98,
-    lastestPrice: 0,
+    latestPrice: 0,
     originalVolume: 529170,
-    lastestVolume: 0,
+    latestVolume: 0,
     change: "0.00",
     "%change": "0.00",
-    status: ""
+    status: "",
   },
   {
     code: "WER.GOW",
     company: "CICB",
     originalPrice: 67.32,
-    lastestPrice: 0,
+    latestPrice: 0,
     originalVolume: 832710,
-    lastestVolume: 0,
+    latestVolume: 0,
     change: "0.00",
     "%change": "0.00",
-    status: ""
-  },
-  {
-    code: "ITO.GOW",
-    company: "STATE FARM",
-    originalPrice: 45.12,
-    lastestPrice: 0,
-    originalVolume: 763412,
-    lastestVolume: 0,
-    change: "0.00",
-    "%change": "0.00",
-    status: ""
+    status: "",
   },
   {
     code: "IOT.GOW",
     company: "COMERICA",
     originalPrice: 98.11,
-    lastestPrice: 0,
+    latestPrice: 0,
     originalVolume: 429872,
-    lastestVolume: 0,
+    latestVolume: 0,
     change: "0.00",
     "%change": "0.00",
-    status: ""
+    status: "",
   },
   {
     code: "RET.GOW",
     company: "E-TRADE",
     originalPrice: 57.12,
-    lastestPrice: 0,
+    latestPrice: 0,
     originalVolume: 871212,
-    lastestVolume: 0,
+    latestVolume: 0,
     change: "0.00",
     "%change": "0.00",
-    status: ""
+    status: "",
   },
   {
     code: "KIG.GOW",
     company: "FROST BANK",
     originalPrice: 67.32,
-    lastestPrice: 0,
+    latestPrice: 0,
     originalVolume: 769283,
-    lastestVolume: 0,
+    latestVolume: 0,
     change: "0.00",
     "%change": "0.00",
-    status: ""
+    status: "",
   },
   {
     code: "DFT.GOW",
     company: "BANK OF HOPE",
     originalPrice: 76.12,
-    lastestPrice: 0,
+    latestPrice: 0,
     originalVolume: 346236,
-    lastestVolume: 0,
+    latestVolume: 0,
     change: "0.00",
     "%change": "0.00",
-    status: ""
+    status: "",
   },
   {
     code: "RPO.GOW",
     company: "WASHINGTON FEDERAL",
     originalPrice: 48.98,
-    lastestPrice: 0,
+    latestPrice: 0,
     originalVolume: 786123,
-    lastestVolume: 0,
+    latestVolume: 0,
     change: "0.00",
     "%change": "0.00",
-    status: ""
+    status: "",
   },
   {
     code: "CVF.GOW",
     company: "ARVEST BANK",
     originalPrice: 68.31,
-    lastestPrice: 0,
+    latestPrice: 0,
     originalVolume: 983722,
-    lastestVolume: 0,
+    latestVolume: 0,
     change: "0.00",
     "%change": "0.00",
-    status: ""
+    status: "",
   },
   {
     code: "DSA.GOW",
     company: "RAYMOND JAMES FINANCIAL",
     originalPrice: 49.09,
-    lastestPrice: 0,
+    latestPrice: 0,
     originalVolume: 467293,
-    lastestVolume: 0,
+    latestVolume: 0,
     change: "0.00",
     "%change": "0.00",
-    status: ""
+    status: "",
   },
   {
     code: "APO.GOW",
     company: "INVESTOR BANK",
     originalPrice: 87.32,
-    lastestPrice: 0,
+    latestPrice: 0,
     originalVolume: 499012,
-    lastestVolume: 0,
+    latestVolume: 0,
     change: "0.00",
     "%change": "0.00",
-    status: ""
+    status: "",
   },
   {
     code: "OIR.GOW",
     company: "SANTANDER BANK",
     originalPrice: 65.23,
-    lastestPrice: 0,
+    latestPrice: 0,
     originalVolume: 459871,
-    lastestVolume: 0,
+    latestVolume: 0,
     change: "0.00",
     "%change": "0.00",
-    status: ""
+    status: "",
   },
   {
     code: "MUT.GOW",
     company: "JFK BANK",
     originalPrice: 68.21,
-    lastestPrice: 0,
+    latestPrice: 0,
     originalVolume: 429847,
-    lastestVolume: 0,
+    latestVolume: 0,
     change: "0.00",
     "%change": "0.00",
-    status: ""
-  }
+    status: "",
+  },
 ];
-
-module.exports = originalData;
